Further work on generating events

+++ src/main/java/uk/gov/justice/digital/hmpps/interventionscatalogue/repository/InterventionSubTypeRepository.java
@@ -4,10 +4,9 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.repository.history.RevisionRepository;
 import org.springframework.stereotype.Repository;
 import uk.gov.justice.digital.hmpps.interventionscatalogue.model.InterventionSubType;
-import uk.gov.justice.digital.hmpps.interventionscatalogue.model.Provider;
 
 import java.util.UUID;
 
 @Repository
-public interface InterventionSubTypeRepository extends RevisionRepository<InterventionSubType, UUID, Integer>, JpaRepository<InterventionSubType, UUID> {
+public interface InterventionSubTypeRepository extends RevisionRepository<InterventionSubType, UUID, Long>, JpaRepository<InterventionSubType, UUID> {
 }
